Feito conversao de emitente

+++ src/AutomacaoNet/DFe/Base/Cidade.cs
@@ -3,7 +3,7 @@
     public class Cidade
     {
         public string Nome { get; set; }
-        public int CodigoIbge { get; set; }
+        public long CodigoIbge { get; set; }
         public EstadoUF EstadoUf { get; set; }
 
         public string SiglaUF => EstadoUf?.Sigla;
